+++ src/main/java/com/example/oatnote/_config/RestTemplateConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.web.client.RestTemplate;
 
-import com.example.oatnote.exception.AiResponseErrorHandler;
+import com.example.oatnote.memo.service.client.exception.AiResponseErrorHandler;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import lombok.RequiredArgsConstructor;
